Implement do_after delays for the attack_hand proc
The sieving functions were refactored into procs that return lists of each item on the turf that fit any of the types in the typelist parameter.

There's a small bug where the insertion operation will continue if you move in certain circumstances. I'm not sure what causes it. I suspect it's do_after's fault.